Add files via upload
Here is the python code on compressing images. It takes the input as the file path of an image and returns the compressed image. It takes one more input 'n_components'. It defines how many no. of components to be retained while compressing the image. The code will be altered to make it more user-friendly,  in python or java, and will be uploaded soon.